# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Наша CRM",
        "version": "1.0.0"
    },
    "paths": {
        "/api/account/{user_id}": {
            "get": {
                "tags": [
                    "Аккаунт"
                ],
                "summary": "Получить Аккаунт по id пользователя",
                "description": "Вернет Аккаунт по id",
                "operationId": "aee8747b3e43116030914d091f61aec0",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "ID Аккаунта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 12.21
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/account/store": {
            "post": {
                "tags": [
                    "Аккаунт"
                ],
                "summary": "Установить начальное значение счета",
                "description": "Установить начальное значение счета",
                "operationId": "95ca035b93b00957231a7642f2602b30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uid",
                                    "total"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "integer",
                                        "example": 11
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                }
            }
        },
...
    }
}